requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
numpy
pandas
scikit-learn
joblib
cryptography
